All compiler warnings are resolved. The build is clean (make app reports zero warnings) and tests pass.
Root cause
DietCodeWindowController was split across category files (MacWindow+Files.mm, MacWindow+AgentAPI.mm, etc.), but methods were still declared on the primary @interface in MacWindow.hpp and the private extension. Clang expected them in the main @implementation block, which produced ~240 -Wincomplete-implementation / -Wobjc-protocol-method-implementation warnings.

Fix
Added MacWindow+CategoryAPI.hpp — named category interfaces (Tabs, Files, AgentAPI, Language, etc.) matching where methods are implemented.
Slimmed MacWindow.hpp — only core methods (init, showWelcome, toggleSidebar, etc.) and properties.
Slimmed MacWindow+Private.hpp — ivars and properties only; method declarations moved to category API.
Split MacWindow.mm — primary @implementation for lifecycle/core actions; (Core) category for helpers like jumpToLine: and updateWindowTitleAndStatus.
Other warnings fixed
MacControlServer.mm — sign-compare: cast _activeConnections.count to NSInteger.
MacControlRecoveryStore.mm — removed unused preimageHash variable.
MacWindow+AgentAPI.mm — removed duplicate jumpToLine:column: that shadowed the Core implementation.
MacWindow+Language.mm — implemented goToDefinitionClicked: (was declared but missing) using the LSP client.

+++ src/platform/macos/control/MacControlServer.mm
@@ -367,7 +367,7 @@ static const void* kDietCodeReadQueueKey = &kDietCodeReadQueueKey;
                 @synchronized(self) {
                     if (!_isRunning) {
                         close(clientFd);
-                    } else if (_activeConnections.count >= kMaxActiveConnections) {
+                    } else if ((NSInteger)_activeConnections.count >= kMaxActiveConnections) {
                         [self logRuntimeDiagnostic:@"unknown" method:@"connection" phase:@"connection_rejected" ok:NO stringCode:@"connection_limit_exceeded" queue:nil durationMs:-1];
                         close(clientFd);
                     } else {
